feat: Add Terraform infrastructure for PostgreSQL operator
Introduce infrastructure setup using Terraform for deploying PostgreSQL operator
on Kubernetes. This includes definition of Kubernetes resources, custom resource
definitions (CRDs) for PostgreSQL, and necessary configuration variables and outputs
for deployment in the specified namespace. This setup deploys the Postgres Operator
with an associated service account and initializes necessary permissions across
the cluster.